% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_dash.R
\name{auto_dash}
\alias{auto_dash}
\title{Generate a dashboard of `ggplot` objects}
\usage{
auto_dash(
  file_path,
  num_plots = 4,
  dash_model = "gpt-4",
  code_model = "gpt-4",
  num_design_attempts = 1,
  num_code_attempts = 5
)
}
\arguments{
\item{file_path}{The path to your input file. Must be .csv or .xlsx.}

\item{num_plots}{The number of plots in your dashboard. Default is 4.}

\item{dash_model}{The name of the language model to use for designing the dashboard. Default 'gpt-4'.}

\item{code_model}{The name of the language model to use for coding individual plots. Default 'gpt-4'.}

\item{num_design_attempts}{The number of iterations to improve on the dashboard design. Default is 1.}

\item{num_code_attempts}{The maximum number of attempts to code your plot before failing - can take less if no errors are encountered in code generation. Default is 5.}
}
\value{
The list of `ggplot` objects representing the dashboard.
}
\description{
This function automatically generates a dashboard of `ggplot` objects from your input file.
It allows you to specify the number of plots and a custom description of your dashboard.
}
\examples{
\dontrun{
# Example usage with saving ggplots to a pdf
all_plots <- auto_dash("iris.csv")
pdf("my_dash.pdf", width = 20, height = 10 * length(all_plots) / 2)
do.call(grid.arrange, c(all_plots, ncol = 2))
dev.off()
}

}
