% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_dash_design.R
\name{auto_dash_design}
\alias{auto_dash_design}
\title{Generate a dashboard design.}
\usage{
auto_dash_design(
  data,
  summary = NULL,
  num_plots = 6,
  custom_description = "",
  dash_model = "gpt-4",
  temperature = 0.1,
  num_design_attempts = 2,
  max_cols = 10,
  filter_pk_cols = FALSE,
  save_messages = TRUE,
  save_dir = "sullyplot_messages",
  save_name = "auto_dash"
)
}
\arguments{
\item{data}{The input data to design the dashboard from, can be either the path to a file (must be .csv or .xlsx) or a data frame. Not used if a summary is provided.}

\item{summary}{Optional summary dataframe of the input file. Should be generated using `himunge::make_df_summary`.}

\item{num_plots}{The number of plots in your dashboard. Default is 6.}

\item{custom_description}{An optional description to describe the custom dashboard you want.}

\item{dash_model}{The name of the language model to use for designing the dashboard. Default 'gpt-4'.}

\item{temperature}{The temperature value for the language model designing the dashboard (does not affect code generation). Default is 0.1.}

\item{num_design_attempts}{The number of iterations to improve on the dashboard design. Default is 2.}

\item{max_cols}{The maximum number of columns the LLM can 'see'. If more columns are provided in the file they will be sorted and the columns with the least NA and most information will be selected. Default is 10.}

\item{filter_pk_cols}{Whether to filter out primary key columns, useful if you have a primary key like an id which you don't want plotted. Default is false.}

\item{save_messages}{Whether to save chat messages and responses for each dashboard generation step, useful for finetuning. Default is false.}

\item{save_dir}{The directory to save chat messages in.}

\item{save_name}{The name to save chat messages under (will be suffixed for each step). Default is "auto_dash".}
}
\value{
A dataframe consisting of the columns input_columns - a list of lists of the input columns necessary for each plot; and descriptions - a list of descriptions of each plot.
}
\description{
This function automatically generates a dataframe describing the design of a dashboard for the given file or file summary,
including the input columns necessary and descriptions for each plot in the dashboard.
Useful for modifying designs and generating them with auto_plot.
}
