% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_dash_df.R
\name{auto_dash_df}
\alias{auto_dash_df}
\title{Generate a dashboard of `ggplot` objects}
\usage{
auto_dash_df(
  file_df,
  num_plots = 6,
  custom_description = "",
  dash_model = "gpt-4",
  code_model = "gpt-4",
  num_design_attempts = 1,
  num_code_attempts = 5,
  max_cols = 10,
  save_messages = FALSE,
  save_dir = "",
  save_name = "auto_dash"
)
}
\arguments{
\item{file_df}{Your input dataframe to be plotted.}

\item{num_plots}{The number of plots in your dashboard. Default is 6.}

\item{custom_description}{An optional description to describe the custom dashboard you want.}

\item{dash_model}{The name of the language model to use for designing the dashboard. Default 'gpt-4'.}

\item{code_model}{The name of the language model to use for coding individual plots. Default 'gpt-4'.}

\item{num_design_attempts}{The number of iterations to improve on the dashboard design. Default is 1.}

\item{num_code_attempts}{The maximum number of attempts to code your plot before failing - can take less if no errors are encountered in code generation. Default is 5.}

\item{max_cols}{The maximum number of columns the LLM can 'see'. If more columns are provided in the file they will be sorted and the columns with the least NA and most information will be selected. Default is 10.}

\item{save_messages}{Whether to save chat messages and responses for each dashboard generation step, useful for finetuning. Default is false.}

\item{save_dir}{The directory to save chat messages in.}

\item{save_name}{The name to save chat messages under (will be suffixed for each step). Default is "auto_dash".}
}
\value{
The list of `ggplot` objects representing the dashboard.
}
\description{
This function automatically generates a dashboard of `ggplot` objects from your input data frame.
It allows you to specify the number of plots and a custom description of your dashboard.
}
