% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{autoconvert_dataframe}
\alias{autoconvert_dataframe}
\title{autoconvert_dataframe}
\usage{
autoconvert_dataframe(df, exclusions = NULL)
}
\arguments{
\item{df}{: The dataframe to convert.}

\item{exclusions}{: columns to explicity exclude from the auto-conversion}
}
\value{
: a named list with two elements: `df` is the input dataframe with all conversions applied and `formats` is a vector of detected formats for the columns of `df`
}
\description{
Detect and apply data converters to a dataframe. Specifically will analyse all character columns and see if they match known formats, and convert them if they do.
This function is useful in an EDA, but be cautious deploying it in production, since it makes assumptions on the fly.
}
