% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{date_converters}
\alias{date_converters}
\title{date_converters}
\usage{
date_converters()
}
\value{
: named list of <is_match, method> pairs
}
\description{
Returns all the converter objects for date formats
}
