% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{get_all_converters}
\alias{get_all_converters}
\title{get_all_converters}
\usage{
get_all_converters()
}
\value{
: a named list of converter-lists, each of which has an is_match function and a method function for conversion
}
\description{
get all converter functions to be used in summary stats & tidy dfs
}
