% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{mi_matrix}
\alias{mi_matrix}
\title{mi_matrix
Makes a mutual information matrix from the numeric columns in the input df.}
\usage{
mi_matrix(df)
}
\arguments{
\item{df}{- the data frame to be analysed}
}
\value{
A string representation of the the mutual information matrix of all numeric columns in the input df
which can be directly used for prompting.
}
\description{
mi_matrix
Makes a mutual information matrix from the numeric columns in the input df.
}
