% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{numeric_converters}
\alias{numeric_converters}
\title{numeric_converters}
\usage{
numeric_converters()
}
\value{
: Named list. Each element of the list is another list with names (is_match, method). The former is a predicate function that determines if a given vector matches the target type
and the latter is a converter function that converts the vector to a numeric type according to a specific pattern.
}
\description{
Get checker & converter functions for numeric types
}
