% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_dash_html.R
\name{render_dash_html}
\alias{render_dash_html}
\title{Renders an interactive dashboard in html from a list of `ggplot` objects}
\usage{
render_dash_html(all_plots, filename = NULL, display = FALSE)
}
\arguments{
\item{all_plots}{The list of `ggplot` objects to render as an interactive dashboard.}

\item{filename}{The filename to save the html file under (must be .html), no file is saved if null. Default is null.}

\item{display}{If true and a filename is specified opens up the dashboard html file in a separate browser page. Default is false}
}
\value{
The html code for the interactive dashboard.
}
\description{
This function renders an interactive dashboard in html given a list of `ggplot` objects, which
can be created using `auto_dash`, `auto_plot`, or created separately from this package.
}
\examples{
\dontrun{
   # Example usage converting plots from `auto_dash` to a pdf dashboard
   output_dir <- tempdir() # Define output directory for saving the dashboard
   all_plots <- sullyplot::auto_dash(system.file("examples/iris.csv", package = "sullyplot"))
   sullyplot::render_dash_html(all_plots, filename=file.path(output_dir, "my_dash.html"), display=TRUE)

   # Note: In actual use, replace tempdir() with a specific directory
   # and ensure the directory exists or is created before saving.
}

}
