% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_dash_pdf.R
\name{render_dash_pdf}
\alias{render_dash_pdf}
\title{Renders an interactive dashboard in html from a list of `ggplot` objects}
\usage{
render_dash_pdf(all_plots, filename)
}
\arguments{
\item{all_plots}{The list of `ggplot` objects to render as an interactive dashboard.}

\item{filename}{The filename to save the pdf file under (must be .pdf).}
}
\value{
The html code for the interactive dashboard.
}
\description{
This function renders an interactive dashboard in html given a list of `ggplot` objects, which
can be created using `auto_dash`, `auto_plot`, or created separately from this package.
}
\examples{
\donttest{
# Example usage converting plots from `auto_dash` to a pdf dashboard
output_dir <- tempdir() # Define output directory for saving the dashboard
all_plots <- sullyplot::auto_dash(system.file("examples/iris.csv", package = "sullyplot"))
sullyplot::render_dash_pdf(all_plots, filename=file.path(output_dir, "my_dash.pdf"))
}

}
