% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_dashboard.R
\name{render_dashboard}
\alias{render_dashboard}
\title{Renders an interactive dashboard from a list of `ggplot` objects}
\usage{
render_dashboard(all_plots, display = FALSE)
}
\arguments{
\item{all_plots}{The list of `ggplot` objects to render as an interactive dashboard.}

\item{display}{If true opens up the dashboard in a separate browser page. Default is false.}
}
\value{
The html code for the interactive dashboard.
}
\description{
This function renders an interactive dashboard given a list of `ggplot` objects, which
can be created using `auto_dash`, `auto_plot`, or created separately from this package.
}
\examples{
\dontrun{
# Example usage converting plots from `auto_dash` to an interactive dashboard
all_plots <- auto_dash("iris.csv")
html_code <- render_dashboard(all_plots, display=TRUE)
}

}
