% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plot_html.R
\name{render_plot_html}
\alias{render_plot_html}
\title{Renders an a `ggplot` object as an interactive html plot.}
\usage{
render_plot_html(plot, filename = NULL, display = FALSE)
}
\arguments{
\item{plot}{The `ggplot` object to render as an interactive html plot.}

\item{filename}{The filename to save the html file under (must be .html), no file is saved if null. Default is null.}

\item{display}{If true and a filename is specified opens up the dashboard html file in a separate browser page. Default is false.}
}
\value{
The html code for the interactive dashboard.
}
\description{
This function renders a `ggplot` object as an interactive plotly-based html plot.
}
