% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{replace_char_pseudonulls}
\alias{replace_char_pseudonulls}
\title{replace_char_pseudonulls
Replaces pseudonulls with NA in the type-char columns of a data frame}
\usage{
replace_char_pseudonulls(df)
}
\arguments{
\item{df}{- the data frame to be analyzed}
}
\value{
the original data frame with the char columns transformed so that pseudonulls are replaced with nulls
}
\description{
replace_char_pseudonulls
Replaces pseudonulls with NA in the type-char columns of a data frame
}
