% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{score_vector_information}
\alias{score_vector_information}
\title{score_vector_information
Assign a score to a vector that is low if the vector has lots of repeats
and higher if the vector has more variation.
The score is the entropy of the histogram (counts) of the vector}
\usage{
score_vector_information(vec)
}
\arguments{
\item{vec}{vector of characters or factors. Not really intended for numeric data.}
}
\value{
A single number, the score of the vector
}
\description{
score_vector_information
Assign a score to a vector that is low if the vector has lots of repeats
and higher if the vector has more variation.
The score is the entropy of the histogram (counts) of the vector
}
