% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{significant_categorical_numeric_relationships}
\alias{significant_categorical_numeric_relationships}
\title{significant_categorical_numeric_relationships
Uses an ANOVA test (or just a t-test if there are two levels) to summarise all significant relationships
between categorical and numeric variables.}
\usage{
significant_categorical_numeric_relationships(
  df,
  summary_df,
  significance_level = 0.05
)
}
\arguments{
\item{df}{- the data frame to be analysed}

\item{summary_df}{- a summary of `df` generated using `summarise_df`}

\item{significance_level}{- the significance level above which relationships will discarded.}
}
\value{
Returns a string summarising significant categorical to numeric relationships and their p values
which can be directly used in prompting.
}
\description{
significant_categorical_numeric_relationships
Uses an ANOVA test (or just a t-test if there are two levels) to summarise all significant relationships
between categorical and numeric variables.
}
