% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{significant_categorical_relationships}
\alias{significant_categorical_relationships}
\title{significant_categorical_relationships
Uses a chi squared test to summarise significant categorical to categorical relationships in the input df.}
\usage{
significant_categorical_relationships(
  df,
  summary_df,
  significance_level = 0.05
)
}
\arguments{
\item{df}{- the data frame to be analysed}

\item{summary_df}{- a summary of `df` generated using `summarise_df`}

\item{significance_level}{- the significance level above which relationships will discarded.}
}
\value{
Returns a string summarising significant categorical to categorical relationships and their p values
which can be directly used in prompting.
}
\description{
significant_categorical_relationships
Uses a chi squared test to summarise significant categorical to categorical relationships in the input df.
}
