% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-package.R
\docType{package}
\name{template}
\alias{template}
\title{template: Name of your module}
\description{
Description goes here. Please rename "template" to your package name EVERYWHERE.
}
\section{template functions}{

\subsection{Section 1}{
\itemize{
\item \code{function_1} - Description 1
\item \code{function_2} - Description 2
}
}
}

\section{datasets}{

\itemize{
\item \code{dataset_1} - description of dataset
}
}

