% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plot.R
\name{auto_plot}
\alias{auto_plot}
\title{Generate code and a `ggplot` object according to the described plot.}
\usage{
auto_plot(
  file_df,
  plot_columns,
  plot_description,
  num_code_attempts = 5,
  code_model = "gpt-4"
)
}
\arguments{
\item{file_df}{A dataframe which you want plotted.}

\item{plot_columns}{The list of names of columns in your dataframe which are relevant to the plot.}

\item{plot_description}{The description of the plot you want.}

\item{num_code_attempts}{The maximum number of attempts to code your plot before failing - can take less if no errors are encountered in code generation. Default is 5.}

\item{code_model}{The name of the language model to use for coding individual plots. Default 'gpt-4'.}
}
\value{
The list of `ggplot` objects representing the dashboard.
}
\description{
This function automatically generates code and a `ggplot` object according to the described plot.
}
\examples{
\dontrun{
# Example usage with saving ggplots to a pdf
my_plot <- auto_plot(iris_df, ["variety", "sepal.length"], "A box plot of sepal length for each variety to show the distribution of sepal length within each variety.", )
print(my_plot)
}
}
