% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sullyplot-package.R
\docType{package}
\name{sullyplot}
\alias{sullyplot}
\title{sullyplot: Package for automatically plotting R plots and dashboards using LLMs.}
\description{
The \code{sullyplot} R Package provides a framework for automated plotting of graphs and dashboards using OpenAI's latest LLMs.
}
\section{sullyplot functions}{

\subsection{Functions}{
\subsection{Automatically plot a full dashboard}{
\enumerate{
\item \code{auto_dash_interactive} - Generates an interactive dashboard from the input file and returns the html.
\item \code{auto_dash} - Generates a dashboard from the input file and returns the list of \code{ggplot} objects.
\item \code{render_dashboard} - Renders a list of \code{ggplot} objects as an interactive dashboard in html.
}
}

\subsection{Automatically plot individual graphs}{
\enumerate{
\item \code{auto_plot} - Generates a \code{ggplot} object from an input file, list of necessary columns, and plot description.
\item \code{attempt_code} - Attempts to plot the input file given the plotting code and returns the resulting status and plot object.
}
}

}
}

